requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.14.0
fsspec==2024.5.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
PyAudio==0.2.14
python-dotenv==1.0.1
regex==2024.5.15
requests==2.31.0
SpeechRecognition==3.10.4
sympy==1.12
tbb==2021.12.0
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
